Proper prototype for TARGA-loading
git-svn-id: https://svn.blender.org/svnroot/verse/trunk@3385 49c17534-1d02-dc11-b283-0030488c597c

+++ la_particle_fx.c
@@ -5,8 +5,8 @@
 
 //#include "presuade.h"
 
-extern boolean	la_load_targa(char *file_name, uint8 *texture_id);
-extern uint		la_save_targa(char *file_name, float *data, unsigned int size);
+extern boolean	la_load_targa(char *file_name, uint *texture_id);
+extern uint	la_save_targa(char *file_name, float *data, unsigned int size);
 
 #define TEX_SPLIT 8
